Added refreshResources ( [ bool refreshAll ] ) to allow resources to update the resource list.
This is useful for refreshing the resources list remotely, i.e. through IRC or webinterface. Currently the only way to refresh the list is by going in-game. But as this is not always possible - or practical - a scripting function for it would be useful.

refreshAll defaults to false, which means this function does the same as /refresh and only loads new resources and unloads removed ones. When set to true, this function does the same as /refreshall, which tends to lag quite a bit, and also checks for changes in the scripts etc.