Added setElementFrozen and isElementFrozen, both client-side and server-side. This replaces 3 other functions:
- isPedFrozen / setPedFrozen
- isVehicleFrozen / setVehicleFrozen
- isObjectStatic / setObjectStatic

Note: It's now possible to make objects static server-side, using this function as well. Previously this was client-side only

Also added a "frozen" setting for map files, where objects, peds and vehicles can be pre-defined as frozen. For instance:

<object model="1305" interior="0" dimension="0" posX="-711.96221923828" posY="964.61572265625" posZ="12.742193222046" rotX="0" rotY="0" rotZ="0" frozen="true" />

This will create a rock near the "Play" spawnpoint, which you'd normally be able to push. But since it's "frozen" (static) it won't move at all.