[Eir] basemodelinfo patch! + misc - added vehicle/clump/atomic model info variations - added streaming functions + custom resource loader - added IMG archive patch (inspired by fastman92) - added GTA:SA pool class + direct access to game pools - improved task interface management (now using global C array for look-up) - added some virtual interfacing for game classes (task, entity) - added transformation class for CPlaceableSAInterface matrix management - added MTA:Eir FileSystem - added building streamer patch - added game TexDictionary management classes (CTxdInstanceSA, ppTxdPool) - removed vehicle LOD + added vehicle rendering functions - updated RenderWare headers + functions - collisions should be more stable now - prepared movement of virtual RenderWare interfaces (solution should be first dicussed with MTA developers) - added memory debugger for corruption and leak detection - updated RenderWare list definitions - added OpenGlobalStream for game access zone management (currently game should access its own directory only) - renamed RwMatrix members to match CMatrix; avoids confusion prolly - improved CPedIntelligenceSAInterface members - updated deathmatch model logic to standardize new colmodel loading interface (no more wild collision loading by clump reading) - updated CAnimBlockSAInterface definition (member functions) - (experimental) increased entity maximum: peds -> 500, vehicles -> 500, objects -> 4000 - reorganized CGameSA and StdInc.h; now it has members of the internal classes instead of virtual interfaces (i.e. CFx -> CFxSA) - fixed various bugs/oddities with model loading (invalid animation load attempt -> infinite loop) - improved CRenderWareSA::ReadClump and CRenderWareSA::ReplaceModel - added various hooks for investigation (CModelInfoSA.ide.cpp, ...)
This patch is huge and I want you to read through all of the changes. :) It can be considered as an easter present: Happy Easter everyone! The bunny came after all.
Though we are not done at this yet; there is much more to improve. First tell me if you like my way of solving problems (C++ native rewrites, removal of ASM code in favor of C++, exposing GTA:SA logic, general performance improvements).

That is all for now. I need time to study for my finals. :/
Meet me at #mta.dev

We are enabling nightly builds for 2.0/Eir soon. Please test it as much as possible. Then report the bugs; you could message me at IRC so I can directly fix them. Or post the bugs at Mantis; will need to get familiar with it.

I tried to keep the changes as little as possible, but adaptions had to be made. I also added documentation to my code additions which you can read to get an understanding of the GTA:SA engine.