[Eir]: Phase 2 - Implementation Update - updated CStreaming environment -> added fibered loading feature -> added alternative data sources features (i.e. cached clothes or complete .img files) -> improved strength of definitions + readability -> split up logic into separate files for easier maintenance -> fixed loading of .IMG files (CMissingModelInfoSA issue fixed) -> added sector streaming logic (responsible for streaming in world entities) - added camera methods - added global player info methods (CPlayerInfoSA.cpp) - added game cache logic (CINFO.BIN management of PS2, disabled by R* to improve mod-ability of GTA:SA) - improved RenderWare interface access (now through global namespace method instead of variable) - fixed any NULL pointer exceptions in TXD management code of GTA:SA - added RenderWare virtual interfaces (in preparation for DFF streaming fix) - added CExecutiveManagerSA asynchronous execution logic (fibers, threads, etc) - added GTA:SA quad tree management logic - added GTA:SA Sectorizer logic - added GTA:SA PathNode loading logic - tons of bugs fixes
This is a big update that pushes code to the Eir fork. It gets the CStreaming implementation up-to-date like it is found in the MTA:Eir Assembla repository. Since the CStreaming implementation depends on a lot of things, I had to include a bunch of stuff as well. This is a code-push revision only. I intend to wire it all up with next commit.

The RenderWare virtual interfaces are what should fix the model support. They are meant to give an internal representation to the game_sa module, so it can ensure integrity of the RenderWare data using virtual classes for mods (i.e. deathmatch module).

So once again a ground-setting commit, sort of like making myself at home (again). Let's go onward!

BUILD STATUS: not-ready