[Eir]: Phase 4 - New Features - added engineStreamingSetProperty (https://wiki.multitheftauto.com/wiki/MTA:Eir/functions/engineStreamingSetProperty) - added engineStreamingGetProperty (https://wiki.multitheftauto.com/wiki/MTA:Eir/functions/engineStreamingGetProperty) - removed streaming optimization from start-up to hint at the new functions; to reproduce the infinite rendering optimization, do: /crun engineStreamingSetProperty( "infiniteStreaming", true ) engineStreamingSetProperty( "nodeStealing", false ) - update to trunk revision 6267
Alright, this marks the beginning of Phase 4, where new features will be added. This is the final Phase of this commit series. I will be taking breaks in this Phase to do University work. I welcome you to host your own Eir fork server and test the new features that will be added in this Phase.

New in this commit is the ability to change streaming behavior. This is a very internal GTA:SA feature that control performance of entity management on your server.
- for race maps, it is good enough to choose a sparse entity management using "gcOnDemand" set to true
- for custom object maps with high farclip, it is best to choose an infinite rendering approach using "infiniteStreaming" set to true and "strictNodeDistrib" set to false; this ensures no rendering artifacts

See for yourself how the streaming modes operate for you! Each has their unique attributes. I will be creating a video to showcase what they are about.