Revert "Add set/getCameraFieldOfView"
This reverts commit 8c62713de880a1676ecb347fb18e432c4145a8aa.

This is because it doesn't work as intended whilst aiming, a potential way
to move forward would be to allow this function to take two arguments: one
regular field of view and another for the field of view whilst aiming.

Right now, when you aim, the field of view returns to how it behaves
normally.