Add new set/getCameraFieldOfView
boolean setCameraFieldOfView(string mode, number value)
number getCameraFieldOfView(string mode)

modes: player/vehicle/vehicle_max

This commit also introduces the requirement to reconnect to the server
you're currently on to apply the field of view settings. This is so that
server-imposed field of view changes cannot be overriden by the client.