Destroy messagebox on disconnect whilst connecting
This commit fixes issue 0008983: 'disconnect' won't destroy the 'Connecting
to...' or 'Entering the game ... window.

Previously it only destroyed (on disconnect) if connecting to the local
server. Now it attempts to destroy the box if any message box is
showing.

This may have side effects but I doubt that any message boxes show whilst
ingame (the browser GUI probably doesn't use this same system).

This commit also makes something DRY.