Fix pasting text when its length is matching gui element's MaxLength
If Editbox or MultiLineEditbox have MaxTextLength set to e.g. 100 characters
then it is impossible to paste any text containing exactly 100 characters
into it.

The best moment to observe this is when user fills the control with text
until it's full, then cuts all of it and tries to paste it again.

Lua test script is available at:
https://github.com/4O4/mtasa-lua-tests/tree/master/%5Btests%5D/test-gui-limited-textbox-paste