[mtaserver.conf] Raise default maxfps to 50
I suggest we change fpslimit to 50 by default, as this is the highest fps avoiding all GTA bugs that might be objections.
- The slower swimming bug isn't present on 50fps, on 60 it is.
- The strafing bug isn't prevent on 50fps.
- The skimmer water take-off bug isn't really present at 50fps (it will just increase its runway to takeoff a bit, but succeed. Acceptable for the benefits of 50fps)

So, 50fps is a breaking point at which we won't get high-fps related GTA bugs. But 50fps feels lot less laggier than default 36, and fact is many servers stick to default value so it will increase overall quality of gameplay within MTA.

Any thoughts? It may be a trivial PR, but the dialogue on whether to do this or not isn't.