Changed triangle FVF to D3DFVF_XYZ
changed it to use D3DFVF_XYZ and use orthagonal projection instead of D3DFVF_XYZRHW to make sure it works properly with render targets, used matrix updates from CTileBatcher.
Made sure that triangles can't be drawn if game is minimized, or if menu is showing. In that case triangle is moved to PreRenderQueue