Implement fast index based lookup for ped pool.
The code is tested and works nicely without breaking anything. This technique allows us to eliminate std::map and looping through arrays by using the index of the object pointer within the pool. Next step is to do the same for other pools.  This will give us a decent performance boost.