Potentional fix for CAnimBlendAssocGroupSA::GetGroupID crash
This crash comes from either CAnimManagerSA::AddAnimation or CAnimManagerSA::ddAnimationAndSync. The problem is that the group ID is out of range (more than 138). When the crash happens, the value of m_pInterface is either 0x50 or 0xA0 of CAnimBlendAssocGroupSA, always. It happens on servers that do not even use IFP animations at all.

The problem with group ID or anim ID corruption already existed, and it has increased during pre-ifp. The crash count is not high for this offset (0x1dc3) on 14334 onwards but nonetheless, this workaround will check the group id, if it's out of range it will log it and switch to idle animation. Once we have the logs, we can use it to reproduce the problem.