Addendum to 9393d581f4c88e28d89781eca4d0b746fd90a7bf (CAnimBlendAssocGroupSA::GetGroupID crash)
The last fix was supposed to check for incorrect group ID. It seems that it had no effect because now we know that the issue was not the group id but the group itself. The group within CAnimManager::ms_aAnimAssocGroups array gets corrupted for some reason, so we should log it and switch to idle animation which is hopefully not corrupt.

In this commit, we are checking for the value of the group pointer in CAnimManager::ms_aAnimAssocGroups array. It will be 0xA0 or 0x50 if it's corrupted, and then the group and animation id will be logged, and we will know exactly when it gets corrupted.