Fix #946: Load right client console input history file using its full filepath
Previously client console input history would not be loaded, or it would load the console.log file which is meant for console outputs and not inputs. To make sure it loads the input file consistently we first fix the input history filename to the one client console uses and then fix to use the calculated path when loading the file (previously it used the relative path).