Revert e0fa528 and 01a1784 (bugfix for the former)
The fix for issue #465 "bone position update delay in onClientPreRender" will be temporarily reverted, because it's currently unstable. Saml1er is working on a fix for onClientPreRender bugs such as right now: cams (like firstperson and spectate) are lagging behind the target, breaking many scripts.
This revert is neccesary to make nightlies stable again, and roll out a quick fix. It can be re-added once fix is ready.