Fix DFF plugin related bugs
At this point, all plugins seem to work if you play gta sa with the custom gta3.img. We need 2 callbacks: GetSize and Write for collision plugin in DFF files.

The collision plugin in DFF files is used by vehicles. For now, we're not optimizing vehicle models since we don't have write and getsize callbacks for collision plugin