Set HW Breakpoint to find cause of Misc13 crash
Misc13 is an old function in CMultiplayerSA_CrashFixHacks.cpp which is trying to avert an ancient  animation crash, but since the avertion has no effect, the crash happens anyway, and which is probably why we have so many anim crashes on various offsets.

After getting some new dumps, we got nice information. Animation groups are getting corrupted  (100% confirmed), we need to find out what is corrupting the members of the animation groups.