Ability to ignore repeated textures to atlas to reduce size of TXD
Reduces size by 20% to 60%, depending on which texture we specify to ignore. We'll do this only for models which are bigger in size, like buildings and not for smaller models, like vehicles or peds. The unwrapped texture can be sometimes very big, and we can reduce the atlas size by not adding a texture which is repeated multiple times.

We'll need to prepare a list of one or two textures for every big model (buildings mostly) which should be ignored as the texture might be repeated multiple times resulting in a massive atlas.