Split new IMG into multiple img files
NOT TESTED. If we keep everything in RAM while generating atlases, we might run out of memory or everything might slow down and take too much time to complete. Later, we'll need to merge all IMG files into one when we are done writing all new DFF and TXD files.