Addendum to 2b40205 (Updated BASS libraries)
BASSFLAC and BASSOPUS contain an updated version of libogg in these updates, for performance improvements and to fix crashes (passing NULL to other libraries).
This was intended to be part of yesterday's commit, but lagged behind and forgotten.

This marks the freeze of any BASS library updates for a reasonable amount of time, everything should be ideally set up now.

*Note: for some reason, BASSOPUS has never been on the integrity checklist