Reverted default SD values from 8cc5f19
Even though it's recommended to use these detections, having it as default can slowly build a climate where certain minorities can barely play on any live server.
Example: cloud gaming, where everything is virtualized and so are control presses. We really do care about this 0.02% of the MTA population.

This information is provided in the commit to avoid discouraging the use of SD #31 and #32 after this change.