Fix #1155: engineFreeModel doesn't "free" itself
This issue is actually in engineRequestModel. CClientModelManager keeps
a "cache" for previously allocated models and attempts to reuse them,
but engineRequestModel always creates a new one.

When engineRequestModel creates a new model, the CClientModel attempts
to add itself to CClientModelManager using ::Add, but ::Add ignores the
attempt, because a CClientModel has already been added for that model
ID. (But we're trying to add a different CClientModel!)

Now:

- Debug assert when CClientModelManager::Add is called with a
  pre-existing model ID (and the Model pointers don't match).
- Re-use existing models if they exist.

There's some crazy code smell here imo - we shouldn't be creating our
own CClientModels, CClientModelManager should be creating them for us.

There's also a good opportunity to use smart pointers here.

These issues should be resolved in a future commit.