Add get/setVehicleModelWheelSize (#1644)
* Add get/setVehicleModelFrontWheelSize and get/setVehicleModelRearWheelSize

These functions, related to issue #719, allow setting the front and
rear wheel size of a vehicle model that GTA: SA reads from the vehicle.ide
file. Moreover, in combination with #1641, all the wheels can then be
visually scaled per vehicle.

* Streamline API of new functions

The new API takes into account suggestions made by StrixG, LosFaul,
qaisjp and others. Thanks!

* Fix returning std::variant with std::unordered_map in new function parser

This fix is needed for the previous commit to compile. According to
@sbx320, the original author of the function parser, it seems correct.

* Replace newly introduced NULL by nullptr

* Remove optional arguments from getVehicleModelWheelSize

* Revert 0765cd3e13186012f739fbdb6e901efd76c942f7 and fix setVehicleModelFrontWheelSize instead

I changed the wrong function because of a mistake.