Assert that m_pUpgrades is not null in CClientVehicle::ResetWheelScale
As explained in previous commits, if the software behaves as specified,
checking for m_pUpgrades is not necessary. Instead, check that invariant
is being held via an assertion, which only impacts debug builds. This
may be useful to catch errors if the `CClientVehicle` class is extended
in such a way that a subclass sets m_pUpgrades to nullptr.