Tweak wheel scaling behavior to be more consistent (#1650)
* Tweak wheel scaling behavior to be more consistent

* Handle m_pUpgrades being nullptr in CClientVehicle::ResetWheelScale

m_pUpgrades shouldn't be null unless its allocation during the
constructor fails without throwing an exception, but other functions
in the same file check for null before using m_pUpgrades, so add a
check here too to be safe.

* Revert "Handle m_pUpgrades being nullptr in CClientVehicle::ResetWheelScale"

This reverts commit b0880ad7b1a78e01226516750c009cd832879a9e.

Checking for nullptr here actually is probably overly defensive, as the
new operator that allocates this field may throw an exception. I think
it is better to be a bit offensive here, because if m_pUpgrades becomes
null something is going very wrong.

* Assert that m_pUpgrades is not null in CClientVehicle::ResetWheelScale

As explained in previous commits, if the software behaves as specified,
checking for m_pUpgrades is not necessary. Instead, check that invariant
is being held via an assertion, which only impacts debug builds. This
may be useful to catch errors if the `CClientVehicle` class is extended
in such a way that a subclass sets m_pUpgrades to nullptr.

* Address suggested change by @qaisjp

* Restore m_fWheelScale in CClientVehicle::Create if needed

Restoring m_fWheelScale to fPreviousWheelScale is needed to handle
succesive restreams properly.