Make element creation more dynamic (based on current/default skin name)
Currently (for new CEGUI) this is set to the default skin. This will be changed to current skin once the skin switching works.

Before this change, you would have to change all references to each element type (e.g "TaharezLook/Button") thoughout the code, when wanting to change skin. This makes it dynamic, based on the skin name.