Condiditional assigment of collision for models to prevent crash.
CGameSA::SetupBrokenModel could crash, if model IDs specified in this function were not registered. This can happen if standard IDE files in gta.dat are commented out.
It's better to do it conditionally, only set the collision for models, if they both exist.
This way we can prevent a crash.