Don't restore default handling for server-sided vehicle changed their model on client-side
That commit won't break any backwards compatibility.
Changing the model for vehicles created on the server side resets the handling only if the vehicle is created locally (client - resets client handling, server - resets server handling).
The problem was caused by the client not checking if the vehicle belongs to them (if it is local) and resetting the handling without checking.