Update cgui images
MTA's official font is Inter.

We dropped Helvetica Neue because of its licensing issues and because not many have
the correct version of the font installed which causes a lot of inconsistencies.

We have been using Inter for a long time on our forums and wiki as well, so many
are familiar with the font. The font also is reminiscent of Helvetica Neue, so change
is minimal, barely noticable for most people.

Inter is distributed under Open Font License (OFL) license which is great for
open source projects that want to achieve the same quality as Helvetica or Arial.

Inter also provides a variable font which means it can have virtually any weight and it
should look good. Inter also has a lot of options built in and has quite wide glyph
support for our use-case.

Inter on GitHub: https://github.com/rsms/inter

Inter on Google Fonts: https://fonts.google.com/specimen/Inter

Description from Google Fonts:

Inter is a variable font family carefully crafted & designed for computer screens.

Inter features a tall x-height to aid in readability of mixed-case and lower-case text.
Several OpenType features are provided as well, like contextual alternates that adjusts
punctuation depending on the shape of surrounding glyphs, slashed zero for when you need
to disambiguate "0" from "o", tabular numbers, etc.

The Inter project is led by Rasmus Andersson, a Swedish maker–of–software living in
San Francisco. To contribute, see github.com/rsms/inter

--

Thanks to Rasmus and all contributors for this great font!