Do not reset handling for custom models on a non-local vehicle (PR #2824)
This pull request doesn't break backwards compatibility with non-custom vehicles. This change will allow the clientside to keep server-side applied handlings, if you set the vehicle model to a custom one requested with engineRequestModel.

Since the server-side doesn't support any custom models (yet - there is a pending pull request for this and needs evaluation after this commit), the models coming from the server will always be in the expected range from 400 to 611 and must always reset the handling on the clientside. If the vehicle is local to the client, the handling must reset everytime. If the vehicle is non-local and the new model id is a custom one, the handling will not be reset and any server-side provided handling will remain and it's the duty of the scripter to synchronize that (currently).