Change method to update Rich Presence
I changed this method to GetTickCount() because the function was being called too frequently, which could lead to additional issues. Now it refreshes every 10 seconds, which should be sufficient. Additionally, in the Discord documentation, there is a note about a rate limit of 1 per 15 seconds, but in reality, it can sometimes be less or more. It also introduces minor improvements.