Revert PR #3620 (Read file from path) due to bugs persisting
After testing, it turns out not to be stable yet, it could be reproduced on FFS server before they changed something (on which the exact details they still need to collaborate with Tracer. We can't accept that a PR may break certain servers due to their specific use of texture/PNG implementations). Investigation into this took place in a gated channel within MTA discord, but will ideally move to dev discord should collaboration to get to the bottom of why the issues still aren't resolved, resume.