Update CVehicleNames.cpp
Vehicles with two names: Vehicles with alternative names are listed in order. For example, “Sadler Damaged”, the “Sadler” row has both the damaged version and the normal version names. and The vehicle names are organized alphabetically so that they can be used consistently in the same order on both server and client side. (the confusion in the table has been eliminated by sorting the server and client side from a to z)