onElementDataChange can be cancelled.
onElementDataChange can now be cancelled when a client, setElementData, or removeElementData changes an element data.
This is useful because right now you have to do setElementData to cancel a change, which broadcasts to all players and also triggers the event again. Being able to just cancel the event is the cleanest way to deal with unauthorized data changes.